test: fix import in tests, changes test dir names
