+++ setup.cfg
@@ -51,7 +51,7 @@ deps =
     tf28: tensorflow ~= 2.8.0
     tf211: tensorflow ~= 2.11.0
 commands =
-    pytest tests/tensorflow tests/common
+    pytest tests/tests_tensorflow tests/tests_common
 
 [testenv:py{38,39,310}-torch{17,110,113,200}]
 deps =
@@ -65,16 +65,17 @@ deps =
     protobuf <=3.20
     tensorflow
     tensorflow_datasets
-    torch17: torch ~= 1.7.1
-    torch17: torchvision ~= 0.8.2
-    torch110: torch ~= 1.10.1
-    torch110: torchvision ~= 0.12.2
-    torch113: torch ~= 1.13.1
-    torch113: torchvision ~= 0.14.1
+    torch17: torch == 1.7.1+cpu
+    torch17: torchvision == 0.8.2+cpu
+    torch110: torch == 1.10.1+cpu
+    torch110: torchvision == 0.12.2+cpu
+    torch113: torch == 1.13.1+cpu
+    torch113: torchvision == 0.14.1+cpu
     torch200: torch
     torch200: torchvision
+install_command = pip install --extra-index-url https://download.pytorch.org/whl/cpu {opts} {packages}
 commands =
-    pytest tests/torch tests/common
+    pytest tests/tests_torch tests/tests_common
 
 [testenv:py{38,39,310}-lint]
 deps =
