+++ tests/tests_tensorflow/datasets/test_tf_datahandler.py
@@ -23,11 +23,11 @@
 import tempfile
 
 import pytest
-import torch
+import tensorflow as tf
 
-from oodeel.datasets.torch_data_handler import TorchDataHandler
-from tests.torch import generate_data
-from tests.torch import generate_data_torch
+from oodeel.datasets.tf_data_handler import TFDataHandler
+from tests.tests_tensorflow import generate_data
+from tests.tests_tensorflow import generate_data_tf
 
 
 def test_get_item_length():
@@ -35,11 +35,11 @@ def test_get_item_length():
     num_labels = 10
     samples = 100
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples
     )  # .batch(samples)
 
-    length = TorchDataHandler.get_item_length(data)
+    length = TFDataHandler.get_item_length(data)
     assert length == 2
 
 
@@ -48,11 +48,11 @@ def test_get_feature_shape():
     num_labels = 10
     samples = 100
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples
     )  # .batch(samples)
 
-    shape = TorchDataHandler.get_feature_shape(data, 0)
+    shape = TFDataHandler.get_feature_shape(data, 0)
     assert shape == input_shape
 
 
@@ -61,11 +61,11 @@ def test_get_dataset_length():
     num_labels = 10
     samples = 100
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples
     )  # .batch(samples)
 
-    cardinality = TorchDataHandler.get_dataset_length(data)
+    cardinality = TFDataHandler.get_dataset_length(data)
     assert cardinality == samples
 
 
@@ -74,52 +74,55 @@ def test_get_input_from_dataset_item():
     num_labels = 10
     samples = 100
 
-    data = generate_data_torch(
+    data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples
     )  # .batch(samples)
 
-    tensor = TorchDataHandler.get_input_from_dataset_item(data[0])
+    for datum in data.take(1):
+        tensor = TFDataHandler.get_input_from_dataset_item(datum)
     assert tensor.shape == (32, 32, 3)
 
 
 @pytest.mark.parametrize(
     "dataset_name, train",
     [
-        ("MNIST", True),
-        ("MNIST", False),
+        ("mnist", True),
+        ("mnist", False),
     ],
 )
-def test_load_torchvision(dataset_name, train, erase_after_test=True):
+def test_load_tensorflow_datasets(dataset_name, train):
     DATASET_INFOS = {
-        "MNIST": {
-            "img_shape": (1, 28, 28),
+        "mnist": {
+            "img_shape": (28, 28, 1),
             "num_samples": {"train": 60000, "test": 10000},
-        },
+        }
     }
     ds_infos = DATASET_INFOS[dataset_name]
     split = ["test", "train"][int(train)]
 
-    handler = TorchDataHandler()
+    handler = TFDataHandler()
 
     with tempfile.TemporaryDirectory() as tmpdirname:
         # define dataset
         dataset = handler.load_dataset(
-            dataset_name, load_kwargs=dict(root=tmpdirname, train=train, download=True)
+            dataset_name,
+            load_kwargs=dict(data_dir=tmpdirname, split=split, download=True),
         )
 
         # dummy item
-        dummy_item = dataset[0]
+        for item in dataset.take(1):
+            dummy_item = item
         dummy_keys = list(dummy_item.keys())
         dummy_shapes = [v.shape for v in dummy_item.values()]
 
         # check keys
-        assert dataset.output_keys == dummy_keys == ["input", "label"]
+        assert list(dataset.element_spec.keys()) == dummy_keys == ["image", "label"]
 
         # check output shape
         assert (
-            dataset.output_shapes
+            [dataset.element_spec[key].shape for key in dataset.element_spec.keys()]
             == dummy_shapes
-            == [torch.Size(ds_infos["img_shape"]), torch.Size([])]
+            == [tf.TensorShape(ds_infos["img_shape"]), tf.TensorShape([])]
         )
 
         # check len of dataset
@@ -129,13 +132,13 @@ def test_load_torchvision(dataset_name, train, erase_after_test=True):
 @pytest.mark.parametrize(
     "x_shape, num_samples, num_labels, one_hot",
     [
-        ((3, 32, 32), 100, 10, True),
-        ((1, 16, 16), 200, 2, False),
+        ((32, 32, 3), 100, 10, True),
+        ((16, 16, 1), 200, 2, False),
         ((64,), 1000, 20, False),
     ],
 )
 def test_load_arrays_and_custom(x_shape, num_labels, num_samples, one_hot):
-    handler = TorchDataHandler()
+    handler = TFDataHandler()
 
     # === define datasets ===
     # tuple / dict numpy
@@ -145,28 +148,32 @@ def test_load_arrays_and_custom(x_shape, num_labels, num_samples, one_hot):
     dict_np = {"input": tuple_np[0], "label": tuple_np[1]}
 
     # tuple / dict torch
-    tuple_torch = (torch.Tensor(tuple_np[0]), torch.Tensor(tuple_np[1]))
-    dict_torch = {"input": tuple_torch[0], "label": tuple_torch[1]}
+    tuple_tf = (
+        tf.convert_to_tensor(tuple_np[0]),
+        tf.convert_to_tensor(tuple_np[1]),
+    )
+    dict_tf = {"input": tuple_tf[0], "label": tuple_tf[1]}
 
     # custom dataset (TensorDataset)
-    tensor_ds_torch = generate_data_torch(
+    tensor_ds_tf = generate_data_tf(
         x_shape=x_shape, num_labels=num_labels, samples=num_samples, one_hot=one_hot
     )
 
     # === load datasets ===
-    for dataset_id in [tuple_np, dict_np, tuple_torch, dict_torch, tensor_ds_torch]:
+    for dataset_id in [tuple_np, dict_np, tuple_tf, dict_tf, tensor_ds_tf]:
         ds = handler.load_dataset(dataset_id, keys=["key_a", "key_b"])
 
         # check registered keys, shapes
-        output_keys = ds.output_keys
-        output_shapes = ds.output_shapes
+        output_keys = list(ds.element_spec.keys())
+        output_shapes = [ds.element_spec[key].shape for key in ds.element_spec.keys()]
         assert output_keys == ["key_a", "key_b"]
         assert output_shapes == [
-            torch.Size(x_shape),
-            torch.Size([num_labels] if one_hot else []),
+            tf.TensorShape(x_shape),
+            tf.TensorShape([num_labels] if one_hot else []),
         ]
         # check item keys, shapes
-        dummy_item = ds[0]
+        for item in ds.take(1):
+            dummy_item = item
         assert list(dummy_item.keys()) == output_keys
         assert list(map(lambda x: x.shape, dummy_item.values())) == output_shapes
 
@@ -174,13 +181,13 @@ def test_load_arrays_and_custom(x_shape, num_labels, num_samples, one_hot):
 @pytest.mark.parametrize(
     "x_shape, num_samples, num_labels, one_hot",
     [
-        ((3, 32, 32), 100, 10, True),
-        ((1, 16, 16), 200, 2, False),
+        ((32, 32, 3), 100, 10, True),
+        ((16, 16, 1), 200, 2, False),
         ((64,), 1000, 20, False),
     ],
 )
 def test_data_handler_full_pipeline(x_shape, num_samples, num_labels, one_hot):
-    handler = TorchDataHandler()
+    handler = TFDataHandler()
 
     # define and load dataset
     dataset_id = generate_data(
@@ -188,18 +195,18 @@ def test_data_handler_full_pipeline(x_shape, num_samples, num_labels, one_hot):
     )
     dataset = handler.load_dataset(dataset_id, keys=["input", "label"])
     assert len(dataset) == num_samples
-    assert dataset.output_shapes[0] == torch.Size(x_shape)
-    assert dataset.output_shapes[1] == (
-        torch.Size([num_labels]) if one_hot else torch.Size([])
+    assert dataset.element_spec["input"].shape == tf.TensorShape(x_shape)
+    assert dataset.element_spec["label"].shape == (
+        tf.TensorShape([num_labels]) if one_hot else tf.TensorShape([])
     )
 
     # filter by label
     a_labels = list(range(num_labels // 2))
     b_labels = list(range(num_labels // 2, num_labels))
     dataset_a = handler.filter_by_feature_value(dataset, "label", a_labels)
-    num_samples_a = len(dataset_a)
+    num_samples_a = handler.get_dataset_length(dataset_a)
     dataset_b = handler.filter_by_feature_value(dataset, "label", b_labels)
-    num_samples_b = len(dataset_b)
+    num_samples_b = handler.get_dataset_length(dataset_b)
     assert num_samples == (num_samples_a + num_samples_b)
 
     # assign feature, map, get feature
@@ -213,24 +220,30 @@ def test_data_handler_full_pipeline(x_shape, num_samples, num_labels, one_hot):
 
     dataset_a = handler.assign_feature_value(dataset_a, "new_feature", 0)
     dataset_a = dataset_a.map(map_fn_a)
-    features_a = torch.Tensor(handler.get_feature_from_ds(dataset_a, "new_feature"))
-    assert torch.all(features_a == torch.Tensor([-3] * num_samples_a))
+    features_a = tf.convert_to_tensor(
+        handler.get_feature_from_ds(dataset_a, "new_feature")
+    )
+    assert tf.reduce_all(features_a == tf.convert_to_tensor([-3] * num_samples_a))
 
     dataset_b = handler.assign_feature_value(dataset_b, "new_feature", 1)
     dataset_b = dataset_b.map(map_fn_b)
-    features_b = torch.Tensor(handler.get_feature_from_ds(dataset_b, "new_feature"))
-    assert torch.all(features_b == torch.Tensor([5] * num_samples_b))
+    features_b = tf.convert_to_tensor(
+        handler.get_feature_from_ds(dataset_b, "new_feature")
+    )
+    assert tf.reduce_all(features_b == tf.convert_to_tensor([5] * num_samples_b))
 
     # concatenate two sub datasets
     dataset_c = handler.merge(dataset_a, dataset_b)
-    features_c = torch.Tensor(handler.get_feature_from_ds(dataset_c, "new_feature"))
-    assert torch.all(features_c == torch.cat([features_a, features_b]))
+    features_c = tf.convert_to_tensor(
+        handler.get_feature_from_ds(dataset_c, "new_feature")
+    )
+    assert tf.reduce_all(features_c == tf.concat([features_a, features_b], axis=0))
 
     # prepare dataloader
     loader = handler.prepare_for_training(dataset_c, 64, True)
     batch = next(iter(loader))
-    assert batch[0].shape == torch.Size([64, *x_shape])
+    assert batch[0].shape == tf.TensorShape([64, *x_shape])
     assert batch[1].shape == (
-        torch.Size([64, num_labels]) if one_hot else torch.Size([64])
+        tf.TensorShape([64, num_labels]) if one_hot else tf.TensorShape([64])
     )
-    assert batch[2].shape == torch.Size([64])
+    assert batch[2].shape == tf.TensorShape([64])
