+++ tests/tests_tensorflow/methods/test_odin.py
@@ -20,35 +20,35 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from oodeel.methods import Energy
-from tests.tensorflow import generate_data
-from tests.tensorflow import generate_data_tf
-from tests.tensorflow import generate_model
+from oodeel.methods import ODIN
+from tests.tests_tensorflow import generate_data
+from tests.tests_tensorflow import generate_data_tf
+from tests.tests_tensorflow import generate_model
 
 
-def test_energy():
+def test_odin():
     """
-    Test Energy
+    Test ODIN
     """
     input_shape = (32, 32, 3)
     num_labels = 10
     samples = 100
 
-    data_x, _ = generate_data(
+    data = generate_data(
         x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
-    )
+    )  # .batch(samples)
 
     model = generate_model(input_shape=input_shape, output_shape=num_labels)
 
-    energy = Energy()
-    energy.fit(model)
-    scores = energy.score(data_x)
+    odin = ODIN(temperature=100, noise=0.1)
+    odin.fit(model)
+    scores = odin.score(data)
 
     assert scores.shape == (100,)
 
     data = generate_data_tf(
         x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
     ).batch(samples)
-    scores = energy.score(data)
+    scores = odin.score(data)
 
     assert scores.shape == (100,)
