+++ tests/tests_tensorflow/models/test_keras_feature_extractor.py
@@ -20,61 +20,50 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from oodeel.methods import MLS
-from tests.tensorflow import generate_data
-from tests.tensorflow import generate_data_tf
-from tests.tensorflow import generate_model
+from oodeel.models.keras_feature_extractor import KerasFeatureExtractor
+from tests.tests_tensorflow import almost_equal
+from tests.tests_tensorflow import generate_data_tf
+from tests.tests_tensorflow import generate_model
 
 
-def test_mls():
-    """Test MLS"""
+def test_predict():
     input_shape = (32, 32, 3)
     num_labels = 10
     samples = 100
 
-    data_x, _ = generate_data(
-        x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
-    )  # .batch(samples)
+    data = generate_data_tf(
+        x_shape=input_shape, num_labels=num_labels, samples=samples
+    ).batch(samples // 2)
 
     model = generate_model(input_shape=input_shape, output_shape=num_labels)
 
-    energy = MLS()
-    energy.fit(model)
-    scores = energy.score(data_x)
+    feature_extractor = KerasFeatureExtractor(model, output_layers_id=[-3])
 
-    assert scores.shape == (100,)
+    model_fe = KerasFeatureExtractor(model, output_layers_id=[-1])
 
-    data_x = generate_data_tf(
-        x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
-    ).batch(samples)
+    last_layer = KerasFeatureExtractor(model, output_layers_id=[-1], input_layer_id=-2)
 
-    scores = energy.score(data_x)
+    pred_model = model.predict(data)
+    pred_feature_extractor = feature_extractor.predict(data)
+    pred_model_fe = model_fe.predict(data)
+    # To obtain the exact same result, add these lines:
+    # pred_feature_extractor = tf.data.Dataset.from_tensor_slices(
+    #    pred_feature_extractor
+    # ).batch(samples // 2)
+    pred_last_layer = last_layer.predict(pred_feature_extractor)
 
-    assert scores.shape == (100,)
+    assert almost_equal(pred_model, pred_model_fe)
+    assert almost_equal(pred_model, pred_last_layer)
 
 
-def test_msp():
-    """Test MLS"""
+def test_get_weights():
     input_shape = (32, 32, 3)
     num_labels = 10
-    samples = 100
-
-    data_x, _ = generate_data(
-        x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
-    )  # .batch(samples)
 
     model = generate_model(input_shape=input_shape, output_shape=num_labels)
 
-    energy = MLS(output_activation="softmax")
-    energy.fit(model)
-    scores = energy.score(data_x)
+    model_fe = KerasFeatureExtractor(model, output_layers_id=[-1])
+    W, b = model_fe.get_weights(-2)
 
-    assert scores.shape == (100,)
-
-    data_x = generate_data_tf(
-        x_shape=input_shape, num_labels=num_labels, samples=samples, one_hot=False
-    ).batch(samples)
-
-    scores = energy.score(data_x)
-
-    assert scores.shape == (100,)
+    assert W.shape == (900, 10)
+    assert b.shape == (10,)
