+++ tests/tests_tensorflow/utils/test_tf_operator.py
