+++ tests/tests_torch/methods/test_energy.py
