+++ tests/tests_torch/methods/test_vim.py
