+++ tests/tests_torch/utils/test_torch_operator.py
